tower-config: get rid of all date functionality
there is some sort of complicated bug with ghc 7.6 where a build works
locally (e.g. in the tower-config sandbox, cabal build and cabal install
work fine) but not when used as a dependency (e.g. in the ivory-px4-hw
sandbox, which depends on tower-config and uses it via cabal add-source,
the same exact sources for tower-config will fail to build).

i really dont have time or paitence to track down what is causing this
nonsense, and luckily we do not need any of the utc date functionality
at the moment, so i'm just killing that functionality in the toml
lib and if it comes back to haunt me later i'll deal with it then