Ignore channels that are never used.
Instead of adding to the Tower AST when someone calls channel, signal,
or period, wait until all the handlers have been declared and use the
list of channels they consumed. That list is precisely the keys of the
global channel map.

This change can produce the Tower AST in a different order, which is
harmless.

It may also drop some synchronous channels that were not used by any
handler even if they're used by some emitters. However, the
tower_syncchans field is used only for debugging, so this is also
harmless. Plus I want to optimize away emitters on channels that have no
handlers, though I haven't gotten that to work yet.