Split GeneratedCode into frontend and backend parts. (API break)
Dependencies (towerModule, towerDepends, towerArtifact) are frontend
generic. ThreadCode and MonitorCode are compat-backend specific. Signal
code should be backend-specific, but is not currently, so it's in a
separate SignalCode type until we're ready to change that.

No frontend API changes, but `runTower` is part of the backend API.

Since the FreeRTOS STM32 backend doesn't call runTower directly but uses
Ivory.Tower.Compile, it isn't affected by this commit.

This commit updates tower-aadl to use the new types.

As far as I know there are no other maintained Tower backends.