Update TowerPlatform to use backend API. (API change)
This breaks tower-freertos-stm32, which I'm pushing a minimal kludge fix
to for the moment.

Although this changes the definition of the TowerPlatform type and adds
a type parameter to it, I couldn't find any user code that mentions this
type by name. So this change is source-compatible with all current
users.