tower: external monitors.
Used to model driver wrappers in AADL.

Monitors is probably the right abstraction here, since a wrapper has
multiple handlers.

It may be possible to make this abstraction w.r.t. channels, but not
easily. Handlers with a channel that has two defined endpoints gets
translated differently in an external (AADL) thread.